pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "capyswarm"
version = "0.1.0"
authors = [
    {name = "Cooper Miller"},
]
description = "A swarm of AI agents powered by Scrapybara"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "scrapybara",
    "python-dotenv",
]

[tool.setuptools]
packages = ["capyswarm"]